[opus] tdm: use dtype aliases for addresses (u32_t LDS / u64_t global)
Replace __UINTPTR_TYPE__ in the tdm_desc/tdm_window address paths with the
opus dtype aliases: LDS addr/offset -> u32_t (LDS is 32-bit on gfx1250),
global addr/offset -> u64_t/i64_t (64-bit flat). set_global_addr keeps the
>>32 split; reinterpret_cast<u64_t>(ptr) is width-legal. Pipelines truncate
the LDS smem pointer explicitly ((uint32_t)reinterpret_cast<uintptr_t>) and
the device test mirrors the u32_t LDS / alias types. Pre-existing gmem/smem
pointer round-trips keep __UINTPTR_TYPE__ (genuinely pointer-width).

+++ csrc/opus_gemm/include/gfx1250/opus_gemm_pipeline_a16w16_clusterlaunch_tdm_splitk_ws_gfx1250.cuh
@@ -337,14 +337,14 @@ void gemm_a16w16_clusterlaunch_tdm_splitk_ws_kernel_gfx1250(opus_gemm_cluster_td
 
         if (wave_id == 0) {
             WindowA w;
-            w.make(reinterpret_cast<uintptr_t>(smem_a), kargs.ptr_a, 0,
+            w.make((uint32_t)reinterpret_cast<uintptr_t>(smem_a), kargs.ptr_a, 0,
                    k_extent, (uint32_t)row_extent_a, (uint64_t)stride_a,
                    (uint32_t)gk0, (uint32_t)tile_row);
             w.desc.set_workgroup_mask(mask_a);   // <=1-WG mask -> 0 -> multicast off; multi-WG mask kept for fan-out
             produce(w, slot_a_b, opus::number<1 + T::kNumSlots>{});       // FREE_A
         } else {  // wave_id == 1 -> B
             WindowB w;
-            w.make(reinterpret_cast<uintptr_t>(smem_b), kargs.ptr_b, 0,
+            w.make((uint32_t)reinterpret_cast<uintptr_t>(smem_b), kargs.ptr_b, 0,
                    k_extent, (uint32_t)row_extent_b, (uint64_t)stride_b,
                    (uint32_t)gk0, (uint32_t)tile_col);
             w.desc.set_workgroup_mask(mask_b);   // <=1-WG mask -> 0 -> multicast off; multi-WG mask kept for fan-out
